// src/app/(resources)/(generated)/types/FinancialPeriodSummary.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type FinancialPeriodSummary = {
  /**
   * @type string, uuid
   */
  id: string;
  /**
   * @type string
   */
  startDate: string;
  /**
   * @type string
   */
  endDate: string;
  /**
   * @type string
   */
  label: string;
  /**
   * @type integer
   */
  transactionCount: number;
  /**
   * @type boolean | undefined
   */
  isStored?: boolean;
  /**
   * @type boolean
   */
  isCurrent?: boolean | null;
};